building out the control interfaces using Mit and straight joint commands
